fix: better user keys in settings
