+++ saas/{{cookiecutter.project_name}}/webapp/src/lib/components/forms/fieldParts/fieldLabel.svelte
@@ -4,6 +4,7 @@
 	import { fieldHasErrors } from './fieldError.svelte';
 	import { Label } from 'flowbite-svelte';
 	import FieldRequiredIndicator from './fieldRequiredIndicator.svelte';
+	import { capitalizeFirstLetter } from '$lib/utils/strings';
 
 	export let field: string;
 	export let text = '';
@@ -11,10 +12,6 @@
 
 	const { superform } = getFormContext();
 	const { errors } = formFieldProxy(superform, field);
-
-	function capitalizeFirstLetter(text: string) {
-		return text.charAt(0).toUpperCase() + text.slice(1);
-	}
 </script>
 
 <Label for={field} class="label" color={fieldHasErrors($errors) ? 'red' : 'gray'}>
