+++ saas/{{cookiecutter.project_name}}/webapp/src/lib/utils/strings.ts
@@ -1,0 +1,3 @@
+export function capitalizeFirstLetter(text: string) {
+    return text.charAt(0).toUpperCase() + text.slice(1);
+}
